Add file support
Add support to pass a file to compile it like a interpreter would do.
This happens by passing the file input to the build_frame() function
and write it inside of the main() function.

BUG:
Passing a file wiht a '\n' will litteraly break the line in the main function
and end up in a syntax error.