Change EDITOR, remove unused function
Change the way EDITOR is work, it now reads the environment variable
EDITOR, and save it to a pointer 'ENV_EDITOR'. If ENV_EDITOR is not empty
the content of the intern EDITOR variable is replaces by the environment
one.

Remove a unused function 'new_edit()', which was never used.